Ensure calls to ASTParser.createBindings use the correct unit resolver.
- Fixes #3428
